+++ deployer/default/config/task/config_createadmin.php
@@ -3,22 +3,17 @@
 namespace Deployer;
 
 use Deployer\Exception\Exception;
-use SourceBroker\DeployerExtendedDatabase\Utility\ConsoleUtility;
+use SourceBroker\DeployerExtendedTypo3\Utility\ConsoleUtility;
 
 task('config:createadmin', function () {
-
-    $username = getenv('DET_CONFIG_CREATEADMIN_USERNAME');
+    $username = getenv('DEP_CONFIG_CREATEADMIN_USERNAME');
     if (!$username) {
         $username = ask('Give new admin username');
     }
     if (!$username) {
-        throw new Exception('Username has to be defined (env DEP_ADMIN_USERNAME or answer a question).',
-            1560175164001);
+        throw new Exception('Username has to be defined (env DEP_ADMIN_USERNAME or answer a question).', 1560175164);
     }
-    $password = substr(
-        str_shuffle('1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZabcefghijklmnopqrstuvwxyz'),
-        0,
-        20);
+    $password = substr(str_shuffle('1234567890ABCDEFGHIJKLMNOPQRSTUVWXYZabcefghijklmnopqrstuvwxyz'), 0, 20);
     output()->writeln('<comment>Generated password: "' . $password . '"</comment>');
 
     if (get('current_stage') == get('target_stage')) {
@@ -33,7 +28,7 @@ task('config:createadmin', function () {
             $activeDir = test('[ -e {{deploy_path}}/release ]') ?
                 get('deploy_path') . '/release' :
                 get('deploy_path') . '/current';
-            $verbosity = (new ConsoleUtility())->getVerbosityAsParameter(output());
+            $verbosity = (new ConsoleUtility())->getVerbosityAsParameter();
             run('cd ' . $activeDir . ' && {{bin/php}} {{bin/deployer}} config:createadmin ' . get('target_stage') . ' --username ' . escapeshellarg($username) . ' --password ' . escapeshellarg($password) . ' ' . $verbosity);
             output()->writeln('<info>Admin for instance "' . get('target_stage') . '" has been created.</info>');
         } catch (\Exception $e) {